modified a lot of ui
