+++ app/src/main/res/color/fluent_default_icon_tint.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/ui_fg"/>
+</selector>
