+++ app/src/main/res/layout/item_select_package.xml
@@ -14,7 +14,6 @@
             android:layout_height="30dp"
             android:layout_centerVertical="true"
             android:layout_marginStart="10dp"
-
             tools:src="@drawable/ic_launcher_background" />
 
         <TextView
@@ -29,8 +28,7 @@
             android:layout_marginBottom="45dp"
             android:gravity="center"
             android:textSize="18sp"
-            android:textColor="#000000"
+            android:textColor="@color/ui_fg"
             tools:text="@string/app_name" />
     </RelativeLayout>
-
 </RelativeLayout>
